+++ ChangeLog.md
@@ -1,3 +1,16 @@
 # Changelog for scannedb-ok
 
+## 0.0.2.1
+- space insertion with ANN
+
+## 0.0.3.0
+- introduced functors to fmap over all categorizations and to preserve
+  and access glyph level
+- space insertion is seen as a categorization task
+- new infrastructure for block categorization with access to
+  before-categorized blocks and look ahead
+- new infrastructure for linearization
+- new module Precision.hs  
+- app: new command for measuring precision of inter-word space insertion
+
 ## Unreleased changes